#53 Get available agent WIP

+++ helpdesk-data/src/main/java/hu/schonherz/training/helpdesk/data/entity/ConversationEntity.java
@@ -14,7 +14,7 @@ public class ConversationEntity extends BaseEntity {
 
     private static final long serialVersionUID = 1L;
 
-    private int agentId;
+    private Long agentId;
 
     private String clientId;
